fix: move access code generation to server side
